+++ src/TodoExtended.Web/Data/TaskTemplate.cs
@@ -2,7 +2,7 @@ namespace TodoExtended.Web.Data;
 
 public class TaskTemplate
 {
-    public int Id { get; set; }
+    public Guid Id { get; set; } = Guid.NewGuid();
     public required string Title { get; set; }
     public required string TaskListId { get; set; }
     public required string TaskListName { get; set; }